Send sequence w/ heartbeat
