+++ pkg/discord/heartbeat.go
@@ -2,17 +2,19 @@ package discord
 
 import (
 	"encoding/json"
+	"fmt"
 	"time"
 )
 
-var heartbeatBody = []byte(`{
-	"op": 1,
-	"d": null
-}`)
 
 // Heartbeat ...
 func (c *Client) heartbeat(interval int) {
 	defer c.waitGroup.Done()
+	var heartbeatBody = fmt.Sprintf(`{
+		"op": 1,
+		"d": %d
+	}`, c.seq))
+	
 	for {
 		if c.heartbeatAck == true {
 			c.heartbeatAck = false
